Enhance Linear Descent RTH feature
This PR takes the existing Linear Descent option for RTH and enhances it. Firstly, Linear Descent is now no longer an RTH type. If previously using `AT_LEAST_LINEAR_DESCENT`, you should now just use `AT_LEAST` as the RTH type. Linear Descent is now a separate option that can be used with and other RTH type. To do this, the target altitude of the descent is now the `nav_rth_home_altitude`. In some cases, if flying below the home position, this could even be a linear ascent.

An option has also been added to decide how far away the linear descent starts. This is via the `nav_rth_linear_descent_start_distance` parameter. Current behaviour can be maintained by setting this to `0` [default]. You can also specify a distance from home when the descent begins. This is in metres and be up to 10km away. Of course, this value can be adjusted if required.