Make PWM allocation a bit smarter.
Only assign servo output if the timer is not used for a motor.
Only assign a motor output if the timer is not used for a servo.

Needs configurator update and more testing.

Checking resource on the MATEKH743 on the CLI, with 3 motors, servos were correctly assigned to TIM4 outpus,
instead of TIM5, even with outputs sets to MC_SERVO and MC_MOTOR