Update
Unfortunately the large pilot logo can't be available on SD. There seems to be an issue with the storage on the MAX7456 chip. The last 33 characters of page 2 seem to corrupt on power loss. After uploading the font, all looks fine. After a reboot, the logo looks fine. But after a power cycle, the last 33 characters are corrupted.

Changes:
- cms.h now back to tabs
- moved logo and stats drawing to own functions, to save repetition.
- Made large pilot logo HD only
- Disabled large pilot logo support with BFCOMPAT mode.
- Stats and logos will self centre, no matter the column width.
- All arm screen times will be increased by 3 seconds with a safehome failure message.