Delay disarm save if emergency rearm is possible
The PR pauses the save after disarm if an emergency rearm is possible. This is to stop things like the RX pausing and the ESC restarting if you disarm in flight. If you are still flying, you will have 5 seconds to rearm. After that, the save will happen. The save will not happen if you rearm within the 5 seconds. If you're on the ground. The emergency rearm system will deactivate quickly. So the save will happen soon after landing. OSD messages are shown to show the state of the save process.