Show STOP instead of IDLE if motor stop enabled
On a fixed wing, you can have the "`throttle_idle`" set to 0. That combined with `max_throttle` set to 2000 gets rid of the silly scaling that goes on. If this is the case, the OSD does not show the correct message. It shows IDLE instead of STOP when the throttle is all the way down. This adds a check for if motor stop is enabled.